Moved watcher for endpoints to services code
